Refactor HigherLevelReviews::ContestableIssuesController to use a common base class, update specs to used shared examples
